backend/requirements.txt
# Django core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Security
django-argon2-cffi==21.3.0
argon2-cffi==23.1.0

# Database
psycopg2-binary==2.9.10

# Production server
gunicorn==23.0.0

# Code quality
flake8==7.1.1
ruff==0.8.4

# PDF export (optional)
reportlab==4.2.5

# Environment variables
python-dotenv==1.0.1
